pyproject.toml
[project]
name = "dungeondice-matrix"
version = "1.0.0"
authors = [
  { name="jwizzle", email="info@hossel.net" },
]
description = "Tools to roll some virtual dice"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matrix-nio[e2e]>=0.24.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "simplematrixbotlib>=2.12.3",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-or-later"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/jwizzle/dungeondice-matrix"
Issues = "https://github.com/jwizzle/dungeondice-matrix/issues"

[project.scripts]
dungeondice-matrix = "dungeondice-matrix.main:start_bot"
